Allows the php interpreter to be overridden
In MacOS and home-brew, php is installed in a different location than the system php (/usr/bin/php). It is sometimes desired to use a different php in order to select the same version and install pecl extensions.

The advantage of #!/usr/bin/env php is that it will use whatever `php` executable appears first in the user's $PATH.

The disadvantage of #!/usr/bin/env php is that it will use whatever `php` executable appears first in the user's $PATH.

https://unix.stackexchange.com/a/29620